refactor: move streaming logic out of StacksChainState and into net, where it belongs
