+++ stackslib/src/chainstate/stacks/db/blocks.rs
@@ -60,6 +60,7 @@ use crate::core::mempool::MAXIMUM_MEMPOOL_TX_CHAINING;
 use crate::core::*;
 use crate::cost_estimates::EstimatorError;
 use crate::net::relay::Relayer;
+use crate::net::stream::{BlockStreamData, HeaderStreamData, MicroblockStreamData, Streamer};
 use crate::net::BlocksInvData;
 use crate::net::Error as net_error;
 use crate::net::ExtendedStacksHeader;
@@ -445,309 +446,6 @@ impl StagingMicroblock {
     #[cfg(test)]
     pub fn try_into_microblock(self) -> Result<StacksMicroblock, StagingMicroblock> {
         StacksMicroblock::consensus_deserialize(&mut &self.block_data[..]).map_err(|_e| self)
-    }
-}
-
-impl MicroblockStreamData {
-    fn stream_count<W: Write>(&mut self, fd: &mut W, count: u64) -> Result<u64, Error> {
-        let mut num_written = 0;
-        while self.num_items_ptr < self.num_items_buf.len() && num_written < count {
-            // stream length prefix
-            test_debug!(
-                "Length prefix: try to send {:?} (ptr={})",
-                &self.num_items_buf[self.num_items_ptr..],
-                self.num_items_ptr
-            );
-            let num_sent = match fd.write(&self.num_items_buf[self.num_items_ptr..]) {
-                Ok(0) => {
-                    // done (disconnected)
-                    test_debug!("Length prefix: wrote 0 bytes",);
-                    return Ok(num_written);
-                }
-                Ok(n) => {
-                    self.num_items_ptr += n;
-                    n as u64
-                }
-                Err(e) => {
-                    if e.kind() == io::ErrorKind::Interrupted {
-                        // EINTR; try again
-                        continue;
-                    } else if e.kind() == io::ErrorKind::WouldBlock
-                        || (cfg!(windows) && e.kind() == io::ErrorKind::TimedOut)
-                    {
-                        // blocked
-                        return Ok(num_written);
-                    } else {
-                        return Err(Error::WriteError(e));
-                    }
-                }
-            };
-            num_written += num_sent;
-            test_debug!(
-                "Length prefix: sent {} bytes ({} total)",
-                num_sent,
-                num_written
-            );
-        }
-        Ok(num_written)
-    }
-}
-
-impl StreamCursor {
-    pub fn new_block(index_block_hash: StacksBlockId) -> StreamCursor {
-        StreamCursor::Block(BlockStreamData {
-            index_block_hash: index_block_hash,
-            offset: 0,
-            total_bytes: 0,
-        })
-    }
-
-    pub fn new_microblock_confirmed(
-        chainstate: &StacksChainState,
-        tail_index_microblock_hash: StacksBlockId,
-    ) -> Result<StreamCursor, Error> {
-        // look up parent
-        let mblock_info = StacksChainState::load_staging_microblock_info_indexed(
-            &chainstate.db(),
-            &tail_index_microblock_hash,
-        )?
-        .ok_or(Error::NoSuchBlockError)?;
-
-        let parent_index_block_hash = StacksBlockHeader::make_index_block_hash(
-            &mblock_info.consensus_hash,
-            &mblock_info.anchored_block_hash,
-        );
-
-        // need to send out the consensus_serialize()'ed array length before sending microblocks.
-        // this is exactly what seq tells us, though.
-        let num_items_buf = ((mblock_info.sequence as u32) + 1).to_be_bytes();
-
-        Ok(StreamCursor::Microblocks(MicroblockStreamData {
-            index_block_hash: StacksBlockId([0u8; 32]),
-            rowid: None,
-            offset: 0,
-            total_bytes: 0,
-            microblock_hash: mblock_info.microblock_hash,
-            parent_index_block_hash: parent_index_block_hash,
-            seq: mblock_info.sequence,
-            unconfirmed: false,
-            num_items_buf: num_items_buf,
-            num_items_ptr: 0,
-        }))
-    }
-
-    pub fn new_microblock_unconfirmed(
-        chainstate: &StacksChainState,
-        anchored_index_block_hash: StacksBlockId,
-        seq: u16,
-    ) -> Result<StreamCursor, Error> {
-        let mblock_info = StacksChainState::load_next_descendant_microblock(
-            &chainstate.db(),
-            &anchored_index_block_hash,
-            seq,
-        )?
-        .ok_or(Error::NoSuchBlockError)?;
-
-        Ok(StreamCursor::Microblocks(MicroblockStreamData {
-            index_block_hash: anchored_index_block_hash.clone(),
-            rowid: None,
-            offset: 0,
-            total_bytes: 0,
-            microblock_hash: mblock_info.block_hash(),
-            parent_index_block_hash: anchored_index_block_hash,
-            seq: seq,
-            unconfirmed: true,
-            num_items_buf: [0u8; 4],
-            num_items_ptr: 4, // stops us from trying to send a length prefix
-        }))
-    }
-
-    pub fn new_headers(
-        chainstate: &StacksChainState,
-        tip: &StacksBlockId,
-        num_headers_requested: u32,
-    ) -> Result<StreamCursor, Error> {
-        let header_info = StacksChainState::load_staging_block_info(chainstate.db(), tip)?
-            .ok_or(Error::NoSuchBlockError)?;
-
-        let num_headers = if header_info.height < (num_headers_requested as u64) {
-            header_info.height as u32
-        } else {
-            num_headers_requested
-        };
-
-        test_debug!("Request for {} headers from {}", num_headers, tip);
-
-        Ok(StreamCursor::Headers(HeaderStreamData {
-            index_block_hash: tip.clone(),
-            offset: 0,
-            total_bytes: 0,
-            num_headers: num_headers,
-            header_bytes: None,
-            end_of_stream: false,
-            corked: false,
-        }))
-    }
-
-    pub fn new_tx_stream(
-        tx_query: MemPoolSyncData,
-        max_txs: u64,
-        height: u64,
-        page_id_opt: Option<Txid>,
-    ) -> StreamCursor {
-        let last_randomized_txid = page_id_opt.unwrap_or_else(|| {
-            let random_bytes = rand::thread_rng().gen::<[u8; 32]>();
-            Txid(random_bytes)
-        });
-
-        StreamCursor::MempoolTxs(TxStreamData {
-            tx_query,
-            last_randomized_txid: last_randomized_txid,
-            tx_buf: vec![],
-            tx_buf_ptr: 0,
-            num_txs: 0,
-            max_txs: max_txs,
-            height: height,
-            corked: false,
-        })
-    }
-
-    fn stream_one_byte<W: Write>(fd: &mut W, b: u8) -> Result<u64, Error> {
-        loop {
-            match fd.write(&[b]) {
-                Ok(0) => {
-                    // done (disconnected)
-                    return Ok(0);
-                }
-                Ok(n) => {
-                    return Ok(n as u64);
-                }
-                Err(e) => {
-                    if e.kind() == io::ErrorKind::Interrupted {
-                        // EINTR; try again
-                        continue;
-                    } else if e.kind() == io::ErrorKind::WouldBlock
-                        || (cfg!(windows) && e.kind() == io::ErrorKind::TimedOut)
-                    {
-                        // blocked
-                        return Ok(0);
-                    } else {
-                        return Err(Error::WriteError(e));
-                    }
-                }
-            }
-        }
-    }
-
-    pub fn get_offset(&self) -> u64 {
-        match self {
-            StreamCursor::Block(ref stream) => stream.offset(),
-            StreamCursor::Microblocks(ref stream) => stream.offset(),
-            StreamCursor::Headers(ref stream) => stream.offset(),
-            // no-op for mempool txs
-            StreamCursor::MempoolTxs(..) => 0,
-        }
-    }
-
-    pub fn add_more_bytes(&mut self, nw: u64) {
-        match self {
-            StreamCursor::Block(ref mut stream) => stream.add_bytes(nw),
-            StreamCursor::Microblocks(ref mut stream) => stream.add_bytes(nw),
-            StreamCursor::Headers(ref mut stream) => stream.add_bytes(nw),
-            // no-op fo mempool txs
-            StreamCursor::MempoolTxs(..) => (),
-        }
-    }
-
-    pub fn stream_to<W: Write>(
-        &mut self,
-        mempool: &MemPoolDB,
-        chainstate: &mut StacksChainState,
-        fd: &mut W,
-        count: u64,
-    ) -> Result<u64, Error> {
-        match self {
-            StreamCursor::Microblocks(ref mut stream) => {
-                let mut num_written = 0;
-                if !stream.unconfirmed {
-                    // Confirmed microblocks are represented as a consensus-encoded vector of
-                    // microblocks, in reverse sequence order.
-                    // Write 4-byte length prefix first
-                    num_written += stream.stream_count(fd, count)?;
-                    StacksChainState::stream_microblocks_confirmed(&chainstate, fd, stream, count)
-                        .and_then(|bytes_sent| Ok(bytes_sent + num_written))
-                } else {
-                    StacksChainState::stream_microblocks_unconfirmed(&chainstate, fd, stream, count)
-                        .and_then(|bytes_sent| Ok(bytes_sent + num_written))
-                }
-            }
-            StreamCursor::MempoolTxs(ref mut tx_stream) => mempool.stream_txs(fd, tx_stream, count),
-            StreamCursor::Headers(ref mut stream) => {
-                let mut num_written = 0;
-                if stream.total_bytes == 0 {
-                    test_debug!("Opening header stream");
-                    let byte_written = StreamCursor::stream_one_byte(fd, '[' as u8)?;
-                    num_written += byte_written;
-                    stream.total_bytes += byte_written;
-                }
-                if stream.total_bytes > 0 {
-                    let mut sent = chainstate.stream_headers(fd, stream, count)?;
-
-                    if stream.end_of_stream && !stream.corked {
-                        // end of stream; cork it
-                        test_debug!("Corking header stream");
-                        let byte_written = StreamCursor::stream_one_byte(fd, ']' as u8)?;
-                        if byte_written > 0 {
-                            sent += byte_written;
-                            stream.total_bytes += byte_written;
-                            stream.corked = true;
-                        }
-                    }
-                    num_written += sent;
-                }
-                Ok(num_written)
-            }
-            StreamCursor::Block(ref mut stream) => chainstate.stream_block(fd, stream, count),
-        }
-    }
-}
-
-impl Streamer for StreamCursor {
-    fn offset(&self) -> u64 {
-        self.get_offset()
-    }
-    fn add_bytes(&mut self, nw: u64) {
-        self.add_more_bytes(nw)
-    }
-}
-
-impl Streamer for HeaderStreamData {
-    fn offset(&self) -> u64 {
-        self.offset
-    }
-    fn add_bytes(&mut self, nw: u64) {
-        self.offset += nw;
-        self.total_bytes += nw;
-    }
-}
-
-impl Streamer for BlockStreamData {
-    fn offset(&self) -> u64 {
-        self.offset
-    }
-    fn add_bytes(&mut self, nw: u64) {
-        self.offset += nw;
-        self.total_bytes += nw;
-    }
-}
-
-impl Streamer for MicroblockStreamData {
-    fn offset(&self) -> u64 {
-        self.offset
-    }
-    fn add_bytes(&mut self, nw: u64) {
-        self.offset += nw;
-        self.total_bytes += nw;
     }
 }
 
@@ -3291,7 +2989,14 @@ impl StacksChainState {
 
     /// Stream a single header's data from disk
     /// If this method returns 0, it's because we're EOF on the header and should begin the next.
-    fn stream_one_header<W: Write>(
+    ///
+    /// The data streamed to `fd` is meant to be part of a JSON array.  The header data will be
+    /// encoded as JSON, and a `,` will be written after it if there are more headers to follow.
+    /// The caller is responsible for writing `[` before writing headers, and writing `]` after all
+    /// headers have been written.
+    ///
+    /// Returns the number of bytes written
+    pub fn stream_one_header<W: Write>(
         blocks_conn: &DBConn,
         block_path: &str,
         fd: &mut W,
@@ -3374,9 +3079,10 @@ impl StacksChainState {
     }
 
     /// Stream multiple headers from disk, moving in reverse order from the chain tip back.
+    /// The format will be a JSON array.
     /// Returns total number of bytes written (will be equal to the number of bytes read).
     /// Returns 0 if we run out of headers
-    fn stream_headers<W: Write>(
+    pub fn stream_headers<W: Write>(
         &self,
         fd: &mut W,
         stream: &mut HeaderStreamData,
@@ -3472,7 +3178,7 @@ impl StacksChainState {
 
     /// Stream a single microblock's data from the staging database.
     /// If this method returns 0, it's because we're EOF on the blob.
-    fn stream_one_microblock<W: Write>(
+    pub fn stream_one_microblock<W: Write>(
         blocks_conn: &DBConn,
         fd: &mut W,
         stream: &mut MicroblockStreamData,
@@ -3530,7 +3236,7 @@ impl StacksChainState {
     /// Stream multiple microblocks from staging, moving in reverse order from the stream tail to the stream head.
     /// Returns total number of bytes written (will be equal to the number of bytes read).
     /// Returns 0 if we run out of microblocks in the staging db
-    fn stream_microblocks_confirmed<W: Write>(
+    pub fn stream_microblocks_confirmed<W: Write>(
         chainstate: &StacksChainState,
         fd: &mut W,
         stream: &mut MicroblockStreamData,
@@ -3597,7 +3303,7 @@ impl StacksChainState {
     }
 
     /// Stream block data from the chunk store.
-    fn stream_data_from_chunk_store<W: Write>(
+    pub fn stream_data_from_chunk_store<W: Write>(
         blocks_path: &str,
         fd: &mut W,
         stream: &mut BlockStreamData,
@@ -10452,128 +10158,7 @@ pub mod test {
         }
     }
 
-    fn stream_one_header_to_vec(
-        blocks_conn: &DBConn,
-        blocks_path: &str,
-        stream: &mut StreamCursor,
-        count: u64,
-    ) -> Result<Vec<u8>, chainstate_error> {
-        if let StreamCursor::Headers(ref mut stream) = stream {
-            let mut bytes = vec![];
-            StacksChainState::stream_one_header(blocks_conn, blocks_path, &mut bytes, stream, count)
-                .map(|nr| {
-                    assert_eq!(bytes.len(), nr as usize);
-
-                    // truncate trailing ',' if it exists
-                    let len = bytes.len();
-                    if len > 0 {
-                        if bytes[len - 1] == ',' as u8 {
-                            let _ = bytes.pop();
-                        }
-                    }
-                    bytes
-                })
-        } else {
-            panic!("not a header stream");
-        }
-    }
-
-    fn stream_one_staging_microblock_to_vec(
-        blocks_conn: &DBConn,
-        stream: &mut StreamCursor,
-        count: u64,
-    ) -> Result<Vec<u8>, chainstate_error> {
-        if let StreamCursor::Microblocks(ref mut stream) = stream {
-            let mut bytes = vec![];
-            StacksChainState::stream_one_microblock(blocks_conn, &mut bytes, stream, count).map(
-                |nr| {
-                    assert_eq!(bytes.len(), nr as usize);
-                    bytes
-                },
-            )
-        } else {
-            panic!("not a microblock stream");
-        }
-    }
-
-    fn stream_chunk_to_vec(
-        blocks_path: &str,
-        stream: &mut StreamCursor,
-        count: u64,
-    ) -> Result<Vec<u8>, chainstate_error> {
-        if let StreamCursor::Block(ref mut stream) = stream {
-            let mut bytes = vec![];
-            StacksChainState::stream_data_from_chunk_store(blocks_path, &mut bytes, stream, count)
-                .map(|nr| {
-                    assert_eq!(bytes.len(), nr as usize);
-                    bytes
-                })
-        } else {
-            panic!("not a block stream");
-        }
-    }
-
-    fn stream_headers_to_vec(
-        chainstate: &mut StacksChainState,
-        stream: &mut StreamCursor,
-        count: u64,
-    ) -> Result<Vec<u8>, chainstate_error> {
-        let mempool = MemPoolDB::open_test(
-            chainstate.mainnet,
-            chainstate.chain_id,
-            &chainstate.root_path,
-        )
-        .unwrap();
-        let mut bytes = vec![];
-        stream
-            .stream_to(&mempool, chainstate, &mut bytes, count)
-            .map(|nr| {
-                assert_eq!(bytes.len(), nr as usize);
-                bytes
-            })
-    }
-
-    fn stream_unconfirmed_microblocks_to_vec(
-        chainstate: &mut StacksChainState,
-        stream: &mut StreamCursor,
-        count: u64,
-    ) -> Result<Vec<u8>, chainstate_error> {
-        let mempool = MemPoolDB::open_test(
-            chainstate.mainnet,
-            chainstate.chain_id,
-            &chainstate.root_path,
-        )
-        .unwrap();
-        let mut bytes = vec![];
-        stream
-            .stream_to(&mempool, chainstate, &mut bytes, count)
-            .map(|nr| {
-                assert_eq!(bytes.len(), nr as usize);
-                bytes
-            })
-    }
-
-    fn stream_confirmed_microblocks_to_vec(
-        chainstate: &mut StacksChainState,
-        stream: &mut StreamCursor,
-        count: u64,
-    ) -> Result<Vec<u8>, chainstate_error> {
-        let mempool = MemPoolDB::open_test(
-            chainstate.mainnet,
-            chainstate.chain_id,
-            &chainstate.root_path,
-        )
-        .unwrap();
-        let mut bytes = vec![];
-        stream
-            .stream_to(&mempool, chainstate, &mut bytes, count)
-            .map(|nr| {
-                assert_eq!(bytes.len(), nr as usize);
-                bytes
-            })
-    }
-
-    fn decode_microblock_stream(mblock_bytes: &Vec<u8>) -> Vec<StacksMicroblock> {
+    pub fn decode_microblock_stream(mblock_bytes: &Vec<u8>) -> Vec<StacksMicroblock> {
         // decode stream
         let mut mblock_ptr = mblock_bytes.as_slice();
         let mut mblocks = vec![];
@@ -10599,569 +10184,6 @@ pub mod test {
             }
         }
         mblocks
-    }
-
-    #[test]
-    fn stacks_db_stream_blocks() {
-        let mut chainstate = instantiate_chainstate(false, 0x80000000, function_name!());
-        let privk = StacksPrivateKey::from_hex(
-            "eb05c83546fdd2c79f10f5ad5434a90dd28f7e3acb7c092157aa1bc3656b012c01",
-        )
-        .unwrap();
-
-        let block = make_16k_block(&privk);
-
-        let consensus_hash = ConsensusHash([2u8; 20]);
-        let parent_consensus_hash = ConsensusHash([1u8; 20]);
-        let index_block_header =
-            StacksBlockHeader::make_index_block_hash(&consensus_hash, &block.block_hash());
-
-        // can't stream a non-existant block
-        let mut stream = StreamCursor::new_block(index_block_header.clone());
-        assert!(stream_chunk_to_vec(&chainstate.blocks_path, &mut stream, 123).is_err());
-
-        // stream unmodified
-        let stream_2 = StreamCursor::new_block(index_block_header.clone());
-        assert_eq!(stream, stream_2);
-
-        // store block to staging
-        store_staging_block(
-            &mut chainstate,
-            &consensus_hash,
-            &block,
-            &parent_consensus_hash,
-            1,
-            2,
-        );
-
-        // stream it back
-        let mut all_block_bytes = vec![];
-        loop {
-            let mut next_bytes =
-                stream_chunk_to_vec(&chainstate.blocks_path, &mut stream, 16).unwrap();
-            if next_bytes.len() == 0 {
-                break;
-            }
-            test_debug!(
-                "Got {} more bytes from staging; add to {} total",
-                next_bytes.len(),
-                all_block_bytes.len()
-            );
-            all_block_bytes.append(&mut next_bytes);
-        }
-
-        // should decode back into the block
-        let staging_block = StacksBlock::consensus_deserialize(&mut &all_block_bytes[..]).unwrap();
-        assert_eq!(staging_block, block);
-
-        // accept it
-        set_block_processed(&mut chainstate, &consensus_hash, &block.block_hash(), true);
-
-        // can still stream it
-        let mut stream = StreamCursor::new_block(index_block_header.clone());
-
-        // stream from chunk store
-        let mut all_block_bytes = vec![];
-        loop {
-            let mut next_bytes =
-                stream_chunk_to_vec(&chainstate.blocks_path, &mut stream, 16).unwrap();
-            if next_bytes.len() == 0 {
-                break;
-            }
-            test_debug!(
-                "Got {} more bytes from chunkstore; add to {} total",
-                next_bytes.len(),
-                all_block_bytes.len()
-            );
-            all_block_bytes.append(&mut next_bytes);
-        }
-
-        // should decode back into the block
-        let staging_block = StacksBlock::consensus_deserialize(&mut &all_block_bytes[..]).unwrap();
-        assert_eq!(staging_block, block);
-    }
-
-    #[test]
-    fn stacks_db_stream_headers() {
-        let mut chainstate = instantiate_chainstate(false, 0x80000000, function_name!());
-        let privk = StacksPrivateKey::from_hex(
-            "eb05c83546fdd2c79f10f5ad5434a90dd28f7e3acb7c092157aa1bc3656b012c01",
-        )
-        .unwrap();
-
-        let mut blocks: Vec<StacksBlock> = vec![];
-        let mut blocks_index_hashes: Vec<StacksBlockId> = vec![];
-
-        // make a linear stream
-        for i in 0..32 {
-            let mut block = make_empty_coinbase_block(&privk);
-
-            if i == 0 {
-                block.header.total_work.work = 1;
-                block.header.total_work.burn = 1;
-            }
-            if i > 0 {
-                block.header.parent_block = blocks.get(i - 1).unwrap().block_hash();
-                block.header.total_work.work =
-                    blocks.get(i - 1).unwrap().header.total_work.work + 1;
-                block.header.total_work.burn =
-                    blocks.get(i - 1).unwrap().header.total_work.burn + 1;
-            }
-
-            let consensus_hash = ConsensusHash([((i + 1) as u8); 20]);
-            let parent_consensus_hash = ConsensusHash([(i as u8); 20]);
-
-            store_staging_block(
-                &mut chainstate,
-                &consensus_hash,
-                &block,
-                &parent_consensus_hash,
-                i as u64,
-                i as u64,
-            );
-
-            blocks_index_hashes.push(StacksBlockHeader::make_index_block_hash(
-                &consensus_hash,
-                &block.block_hash(),
-            ));
-            blocks.push(block);
-        }
-
-        let mut blocks_fork = blocks[0..16].to_vec();
-        let mut blocks_fork_index_hashes = blocks_index_hashes[0..16].to_vec();
-
-        // make a stream that branches off
-        for i in 16..32 {
-            let mut block = make_empty_coinbase_block(&privk);
-
-            if i == 16 {
-                block.header.parent_block = blocks.get(i - 1).unwrap().block_hash();
-                block.header.total_work.work =
-                    blocks.get(i - 1).unwrap().header.total_work.work + 1;
-                block.header.total_work.burn =
-                    blocks.get(i - 1).unwrap().header.total_work.burn + 2;
-            } else {
-                block.header.parent_block = blocks_fork.get(i - 1).unwrap().block_hash();
-                block.header.total_work.work =
-                    blocks_fork.get(i - 1).unwrap().header.total_work.work + 1;
-                block.header.total_work.burn =
-                    blocks_fork.get(i - 1).unwrap().header.total_work.burn + 2;
-            }
-
-            let consensus_hash = ConsensusHash([((i + 1) as u8) | 0x80; 20]);
-            let parent_consensus_hash = if i == 16 {
-                ConsensusHash([(i as u8); 20])
-            } else {
-                ConsensusHash([(i as u8) | 0x80; 20])
-            };
-
-            store_staging_block(
-                &mut chainstate,
-                &consensus_hash,
-                &block,
-                &parent_consensus_hash,
-                i as u64,
-                i as u64,
-            );
-
-            blocks_fork_index_hashes.push(StacksBlockHeader::make_index_block_hash(
-                &consensus_hash,
-                &block.block_hash(),
-            ));
-            blocks_fork.push(block);
-        }
-
-        // can't stream a non-existant header
-        assert!(StreamCursor::new_headers(&chainstate, &StacksBlockId([0x11; 32]), 1).is_err());
-
-        // stream back individual headers
-        for i in 0..blocks.len() {
-            let mut stream =
-                StreamCursor::new_headers(&chainstate, &blocks_index_hashes[i], 1).unwrap();
-            let mut next_header_bytes = vec![];
-            loop {
-                // torture test
-                let mut next_bytes = stream_one_header_to_vec(
-                    &chainstate.db(),
-                    &chainstate.blocks_path,
-                    &mut stream,
-                    25,
-                )
-                .unwrap();
-                if next_bytes.len() == 0 {
-                    break;
-                }
-                next_header_bytes.append(&mut next_bytes);
-            }
-            test_debug!("Got {} total bytes", next_header_bytes.len());
-            let header: ExtendedStacksHeader =
-                serde_json::from_reader(&mut &next_header_bytes[..]).unwrap();
-
-            assert_eq!(header.consensus_hash, ConsensusHash([(i + 1) as u8; 20]));
-            assert_eq!(header.header, blocks[i].header);
-
-            if i > 0 {
-                assert_eq!(header.parent_block_id, blocks_index_hashes[i - 1]);
-            }
-        }
-
-        // stream back a run of headers
-        let block_expected_headers: Vec<StacksBlockHeader> =
-            blocks.iter().rev().map(|blk| blk.header.clone()).collect();
-
-        let block_expected_index_hashes: Vec<StacksBlockId> = blocks_index_hashes
-            .iter()
-            .rev()
-            .map(|idx| idx.clone())
-            .collect();
-
-        let block_fork_expected_headers: Vec<StacksBlockHeader> = blocks_fork
-            .iter()
-            .rev()
-            .map(|blk| blk.header.clone())
-            .collect();
-
-        let block_fork_expected_index_hashes: Vec<StacksBlockId> = blocks_fork_index_hashes
-            .iter()
-            .rev()
-            .map(|idx| idx.clone())
-            .collect();
-
-        // get them all -- ask for more than there is
-        let mut stream =
-            StreamCursor::new_headers(&chainstate, blocks_index_hashes.last().unwrap(), 4096)
-                .unwrap();
-        let header_bytes =
-            stream_headers_to_vec(&mut chainstate, &mut stream, 1024 * 1024).unwrap();
-
-        eprintln!(
-            "headers: {}",
-            String::from_utf8(header_bytes.clone()).unwrap()
-        );
-        let headers: Vec<ExtendedStacksHeader> =
-            serde_json::from_reader(&mut &header_bytes[..]).unwrap();
-
-        assert_eq!(headers.len(), block_expected_headers.len());
-        for ((i, h), eh) in headers
-            .iter()
-            .enumerate()
-            .zip(block_expected_headers.iter())
-        {
-            assert_eq!(h.header, *eh);
-            assert_eq!(h.consensus_hash, ConsensusHash([(32 - i) as u8; 20]));
-            if i + 1 < block_expected_index_hashes.len() {
-                assert_eq!(h.parent_block_id, block_expected_index_hashes[i + 1]);
-            }
-        }
-
-        let mut stream =
-            StreamCursor::new_headers(&chainstate, blocks_fork_index_hashes.last().unwrap(), 4096)
-                .unwrap();
-        let header_bytes =
-            stream_headers_to_vec(&mut chainstate, &mut stream, 1024 * 1024).unwrap();
-        let fork_headers: Vec<ExtendedStacksHeader> =
-            serde_json::from_reader(&mut &header_bytes[..]).unwrap();
-
-        assert_eq!(fork_headers.len(), block_fork_expected_headers.len());
-        for ((i, h), eh) in fork_headers
-            .iter()
-            .enumerate()
-            .zip(block_fork_expected_headers.iter())
-        {
-            let consensus_hash = if i >= 16 {
-                ConsensusHash([((32 - i) as u8); 20])
-            } else {
-                ConsensusHash([((32 - i) as u8) | 0x80; 20])
-            };
-
-            assert_eq!(h.header, *eh);
-            assert_eq!(h.consensus_hash, consensus_hash);
-            if i + 1 < block_fork_expected_index_hashes.len() {
-                assert_eq!(h.parent_block_id, block_fork_expected_index_hashes[i + 1]);
-            }
-        }
-
-        assert_eq!(fork_headers[16..32], headers[16..32]);
-
-        // ask for only a few
-        let mut stream =
-            StreamCursor::new_headers(&chainstate, blocks_index_hashes.last().unwrap(), 10)
-                .unwrap();
-        let mut header_bytes = vec![];
-        loop {
-            // torture test
-            let mut next_bytes = stream_headers_to_vec(&mut chainstate, &mut stream, 17).unwrap();
-            if next_bytes.len() == 0 {
-                break;
-            }
-            header_bytes.append(&mut next_bytes);
-        }
-
-        eprintln!(
-            "header bytes: {}",
-            String::from_utf8(header_bytes.clone()).unwrap()
-        );
-
-        let headers: Vec<ExtendedStacksHeader> =
-            serde_json::from_reader(&mut &header_bytes[..]).unwrap();
-
-        assert_eq!(headers.len(), 10);
-        for (i, hdr) in headers.iter().enumerate() {
-            assert_eq!(hdr.header, block_expected_headers[i]);
-            assert_eq!(hdr.parent_block_id, block_expected_index_hashes[i + 1]);
-        }
-
-        // ask for only a few
-        let mut stream =
-            StreamCursor::new_headers(&chainstate, blocks_fork_index_hashes.last().unwrap(), 10)
-                .unwrap();
-        let mut header_bytes = vec![];
-        loop {
-            // torture test
-            let mut next_bytes = stream_headers_to_vec(&mut chainstate, &mut stream, 17).unwrap();
-            if next_bytes.len() == 0 {
-                break;
-            }
-            header_bytes.append(&mut next_bytes);
-        }
-        let headers: Vec<ExtendedStacksHeader> =
-            serde_json::from_reader(&mut &header_bytes[..]).unwrap();
-
-        assert_eq!(headers.len(), 10);
-        for (i, hdr) in headers.iter().enumerate() {
-            assert_eq!(hdr.header, block_fork_expected_headers[i]);
-            assert_eq!(hdr.parent_block_id, block_fork_expected_index_hashes[i + 1]);
-        }
-    }
-
-    #[test]
-    fn stacks_db_stream_staging_microblocks() {
-        let mut chainstate = instantiate_chainstate(false, 0x80000000, function_name!());
-        let privk = StacksPrivateKey::from_hex(
-            "eb05c83546fdd2c79f10f5ad5434a90dd28f7e3acb7c092157aa1bc3656b012c01",
-        )
-        .unwrap();
-
-        let block = make_empty_coinbase_block(&privk);
-        let mut mblocks = make_sample_microblock_stream(&privk, &block.block_hash());
-        mblocks.truncate(15);
-
-        let consensus_hash = ConsensusHash([2u8; 20]);
-        let parent_consensus_hash = ConsensusHash([1u8; 20]);
-        let index_block_header =
-            StacksBlockHeader::make_index_block_hash(&consensus_hash, &block.block_hash());
-
-        // can't stream a non-existant microblock
-        if let Err(super::Error::NoSuchBlockError) =
-            StreamCursor::new_microblock_confirmed(&chainstate, index_block_header.clone())
-        {
-        } else {
-            panic!("Opened nonexistant microblock");
-        }
-
-        if let Err(super::Error::NoSuchBlockError) =
-            StreamCursor::new_microblock_unconfirmed(&chainstate, index_block_header.clone(), 0)
-        {
-        } else {
-            panic!("Opened nonexistant microblock");
-        }
-
-        // store microblocks to staging and stream them back
-        for (i, mblock) in mblocks.iter().enumerate() {
-            store_staging_microblock(
-                &mut chainstate,
-                &consensus_hash,
-                &block.block_hash(),
-                mblock,
-            );
-
-            // read back all the data we have so far, block-by-block
-            let mut staging_mblocks = vec![];
-            for j in 0..(i + 1) {
-                let mut next_mblock_bytes = vec![];
-                let mut stream = StreamCursor::new_microblock_unconfirmed(
-                    &chainstate,
-                    index_block_header.clone(),
-                    j as u16,
-                )
-                .unwrap();
-                loop {
-                    let mut next_bytes =
-                        stream_one_staging_microblock_to_vec(&chainstate.db(), &mut stream, 4096)
-                            .unwrap();
-                    if next_bytes.len() == 0 {
-                        break;
-                    }
-                    test_debug!(
-                        "Got {} more bytes from staging; add to {} total",
-                        next_bytes.len(),
-                        next_mblock_bytes.len()
-                    );
-                    next_mblock_bytes.append(&mut next_bytes);
-                }
-                test_debug!("Got {} total bytes", next_mblock_bytes.len());
-
-                // should deserialize to a microblock
-                let staging_mblock =
-                    StacksMicroblock::consensus_deserialize(&mut &next_mblock_bytes[..]).unwrap();
-                staging_mblocks.push(staging_mblock);
-            }
-
-            assert_eq!(staging_mblocks.len(), mblocks[0..(i + 1)].len());
-            for j in 0..(i + 1) {
-                test_debug!("check {}", j);
-                assert_eq!(staging_mblocks[j], mblocks[j])
-            }
-
-            // can also read partial stream in one shot, from any seq
-            for k in 0..(i + 1) {
-                test_debug!("start at seq {}", k);
-                let mut staging_mblock_bytes = vec![];
-                let mut stream = StreamCursor::new_microblock_unconfirmed(
-                    &chainstate,
-                    index_block_header.clone(),
-                    k as u16,
-                )
-                .unwrap();
-                loop {
-                    let mut next_bytes =
-                        stream_unconfirmed_microblocks_to_vec(&mut chainstate, &mut stream, 4096)
-                            .unwrap();
-                    if next_bytes.len() == 0 {
-                        break;
-                    }
-                    test_debug!(
-                        "Got {} more bytes from staging; add to {} total",
-                        next_bytes.len(),
-                        staging_mblock_bytes.len()
-                    );
-                    staging_mblock_bytes.append(&mut next_bytes);
-                }
-
-                test_debug!("Got {} total bytes", staging_mblock_bytes.len());
-
-                // decode stream
-                let staging_mblocks = decode_microblock_stream(&staging_mblock_bytes);
-
-                assert_eq!(staging_mblocks.len(), mblocks[k..(i + 1)].len());
-                for j in 0..staging_mblocks.len() {
-                    test_debug!("check {}", j);
-                    assert_eq!(staging_mblocks[j], mblocks[k + j])
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn stacks_db_stream_confirmed_microblocks() {
-        let mut chainstate = instantiate_chainstate(false, 0x80000000, function_name!());
-        let privk = StacksPrivateKey::from_hex(
-            "eb05c83546fdd2c79f10f5ad5434a90dd28f7e3acb7c092157aa1bc3656b012c01",
-        )
-        .unwrap();
-
-        let block = make_empty_coinbase_block(&privk);
-        let mut mblocks = make_sample_microblock_stream(&privk, &block.block_hash());
-        mblocks.truncate(5);
-
-        let mut child_block = make_empty_coinbase_block(&privk);
-        child_block.header.parent_block = block.block_hash();
-        child_block.header.parent_microblock = mblocks.last().as_ref().unwrap().block_hash();
-        child_block.header.parent_microblock_sequence =
-            mblocks.last().as_ref().unwrap().header.sequence;
-
-        let consensus_hash = ConsensusHash([2u8; 20]);
-        let parent_consensus_hash = ConsensusHash([1u8; 20]);
-        let child_consensus_hash = ConsensusHash([3u8; 20]);
-
-        let index_block_header =
-            StacksBlockHeader::make_index_block_hash(&consensus_hash, &block.block_hash());
-
-        // store microblocks to staging
-        for (i, mblock) in mblocks.iter().enumerate() {
-            store_staging_microblock(
-                &mut chainstate,
-                &consensus_hash,
-                &block.block_hash(),
-                mblock,
-            );
-        }
-
-        // store block to staging
-        store_staging_block(
-            &mut chainstate,
-            &consensus_hash,
-            &block,
-            &parent_consensus_hash,
-            1,
-            2,
-        );
-
-        // store child block to staging
-        store_staging_block(
-            &mut chainstate,
-            &child_consensus_hash,
-            &child_block,
-            &consensus_hash,
-            1,
-            2,
-        );
-
-        // accept it
-        set_block_processed(&mut chainstate, &consensus_hash, &block.block_hash(), true);
-        set_block_processed(
-            &mut chainstate,
-            &child_consensus_hash,
-            &child_block.block_hash(),
-            true,
-        );
-
-        for i in 0..mblocks.len() {
-            // set different parts of this stream as confirmed
-            set_microblocks_processed(
-                &mut chainstate,
-                &child_consensus_hash,
-                &child_block.block_hash(),
-                &mblocks[i].block_hash(),
-            );
-
-            // verify that we can stream everything
-            let microblock_index_header =
-                StacksBlockHeader::make_index_block_hash(&consensus_hash, &mblocks[i].block_hash());
-            let mut stream = StreamCursor::new_microblock_confirmed(
-                &chainstate,
-                microblock_index_header.clone(),
-            )
-            .unwrap();
-
-            let mut confirmed_mblock_bytes = vec![];
-            loop {
-                let mut next_bytes =
-                    stream_confirmed_microblocks_to_vec(&mut chainstate, &mut stream, 16).unwrap();
-                if next_bytes.len() == 0 {
-                    break;
-                }
-                test_debug!(
-                    "Got {} more bytes from staging; add to {} total",
-                    next_bytes.len(),
-                    confirmed_mblock_bytes.len()
-                );
-                confirmed_mblock_bytes.append(&mut next_bytes);
-            }
-
-            // decode stream (should be length-prefixed)
-            let mut confirmed_mblocks =
-                Vec::<StacksMicroblock>::consensus_deserialize(&mut &confirmed_mblock_bytes[..])
-                    .unwrap();
-
-            confirmed_mblocks.reverse();
-
-            assert_eq!(confirmed_mblocks.len(), mblocks[0..(i + 1)].len());
-            for j in 0..(i + 1) {
-                test_debug!("check {}", j);
-                assert_eq!(confirmed_mblocks[j], mblocks[j])
-            }
-        }
     }
 
     #[test]
